correct type casting and remove unused use statement

+++ includes/globalinclude.php
@@ -27,8 +27,8 @@ defined('DB_NAME') || die('Kein MySql Datenbanknamen angegeben. Bitte config.JTL
 defined('DB_USER') || die('Kein MySql-Datenbank Benutzer angegeben. Bitte config.JTL-Shop.ini.php bearbeiten!');
 defined('DB_PASS') || die('Kein MySql-Datenbank Passwort angegeben. Bitte config.JTL-Shop.ini.php bearbeiten!');
 
-define('JTL_VERSION', sprintf('%d%02d', Version::parse(APPLICATION_VERSION)->getMajor(), Version::parse(APPLICATION_VERSION)->getMinor())); // DEPRECATED since 5.0.0
-define('JTL_MINOR_VERSION',Version::parse(APPLICATION_VERSION)->getPatch()); // DEPRECATED since 5.0.0
+define('JTL_VERSION', (int) sprintf('%d%02d', Version::parse(APPLICATION_VERSION)->getMajor(), Version::parse(APPLICATION_VERSION)->getMinor())); // DEPRECATED since 5.0.0
+define('JTL_MINOR_VERSION', (int) Version::parse(APPLICATION_VERSION)->getPatch()); // DEPRECATED since 5.0.0
 
 Profiler::start();
